chore: Add script to dump Martian DB
This commit adds a new script, `dump_martian_db.py`, which is responsible for dumping the Martian DB. This script is executed as part of the automation process in the `automate.sh` file.

The purpose of this script is to generate questions, answers, and judgments for the project. It is configured using the `scripts/configs/dump_martian_db.yaml` file.

This commit also includes some minor changes to the `automate.sh` file to ensure the script is executed correctly.